[#3] Scaffold test for Monster class
I'm wrapping up for the night, so I want to go ahead and get this work
committed. It's still throwing an error at me related to the
generic-pool module, so I'll need to work out that kink to get things
smooth. I may just remove pooling for now and only use a in-class
connection to the database. We'll see.